ai-service/requirements.txt
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0
